Add files via upload
Function `get_gifNimage()` opens any image (jpg, jpeg, png, svg) or gif from a link (in string format). 
After that, it will be saved in the current folder, convert (if needed) to png (from svg format) and– finally – displayed. 
The function deletes the svg file after conversion – in that case, it will leave only the png version so there won't be any useless files in the folder.